+++ evaluation-microservice/Asincrona/evaluation-microservice/src/main/java/com/example/evaluation/infra/dto/DegreeWorkUpdateDTO.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class DegreeWorkUpdateDTO {
 
-    private Long degreeWorkId; // ID del trabajo de grado
+    private Integer degreeWorkId; // ID del trabajo de grado
     private String estado; // Nuevo estado (EnumEstadoDegreeWork en texto)
     private String correcciones; // Texto con las observaciones o correcciones
 }
